+++ astropy/coordinates/tests/test_sky_coord.py
@@ -708,3 +708,37 @@ def test_units_known_fail():
 def test_nodata_failure():
     with pytest.raises(ValueError):
         SkyCoord()
+
+
+def test_frame_attr_transform_inherit():
+    """
+    Test that frame attributes get inherited as expected during transform.
+    Driven by #3106.
+    """
+    c = SkyCoord(1 * u.deg, 2 * u.deg, frame=FK5)
+    c2 = c.transform_to(FK4)
+    assert c2.equinox.value == 'B1950.000'
+    assert c2.obstime.value == 'B1950.000'
+
+    c2 = c.transform_to(FK4(equinox='J1975', obstime='J1980'))
+    assert c2.equinox.value == 'J1975.000'
+    assert c2.obstime.value == 'J1980.000'
+
+    c = SkyCoord(1 * u.deg, 2 * u.deg, frame=FK4)
+    c2 = c.transform_to(FK5)
+    assert c2.equinox.value == 'J2000.000'
+    assert c2.obstime is None
+
+    c = SkyCoord(1 * u.deg, 2 * u.deg, frame=FK4, obstime='J1980')
+    c2 = c.transform_to(FK5)
+    assert c2.equinox.value == 'J2000.000'
+    assert c2.obstime.value == 'J1980.000'
+
+    c = SkyCoord(1 * u.deg, 2 * u.deg, frame=FK4, equinox='J1975', obstime='J1980')
+    c2 = c.transform_to(FK5)
+    assert c2.equinox.value == 'J1975.000'
+    assert c2.obstime.value == 'J1980.000'
+
+    c2 = c.transform_to(FK5(equinox='J1990'))
+    assert c2.equinox.value == 'J1990.000'
+    assert c2.obstime.value == 'J1980.000'